+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2024-06-25
+
 ### Added
 
 - Add per cluster and installation overridable sharding strategy support for mimir-backed installations.
@@ -60,7 +62,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initialize project and create heartbeat for the installation.
 
-[Unreleased]: https://github.com/giantswarm/observability-operator/compare/v0.1.1...HEAD
+[Unreleased]: https://github.com/giantswarm/observability-operator/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/giantswarm/observability-operator/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/giantswarm/observability-operator/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/giantswarm/observability-operator/compare/v0.0.4...v0.1.0
 [0.0.4]: https://github.com/giantswarm/observability-operator/compare/v0.0.3...v0.0.4